% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{RStudio Addin for annotating plate images}
\usage{
annotate(dir = NULL, queries = getOption("screenmill.queries"),
  strain_collections = getOption("screenmill.strain_collections"),
  strain_collection_keys = getOption("screenmill.strain_collection_keys"),
  strains = getOption("screenmill.strains"),
  media = getOption("screenmill.media"),
  treatments = getOption("screenmill.treatments"), temperatures = c(23, 27,
  30, 33, 37), overwrite = FALSE, update = TRUE)
}
\arguments{
\item{dir}{Directory of images to process. If \code{NULL} the
user will be prompted to choose a directory.}

\item{queries}{A dataframe of available queries. Uses
\code{getOption('screenmill.queries')} by default (see Details).}

\item{strain_collections}{A dataframe of available strain
collections. Uses \code{getOption('screenmill.strain_collections')}
(see Details).}

\item{media}{A dataframe of available media. Uses
\code{getOption('screenmill.media')} by default(see Details).}

\item{treatments}{A dataframe of available treatments. Uses
\code{getOption('screenmill.treatments')} by default (see Details).}

\item{temperatures}{A vector of recommended temperatures.}

\item{overwrite}{Should existing annotations be overwritten?
Defaults to \code{FALSE}.}
}
\description{
This function launches a Shiny gadget that allows a user to record information
about plate images for arrayed colony growth experiments.
}
\details{
This application is provided to make annotation of plates in
colony growth experiments easier and safer. It will try to warn, or prevent
the user from entering invalid data. If the \code{rothfreezer} package
is installed, arguments set to \code{NULL} will be pulled from this package's
database.

The resulting annotation table is saved as \code{plate-annotations.csv}
in the directory containing the annotated images.

To quit without saving, just press "Cancel".
}

