% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-annotate-plates.R
\name{annotate_plates}
\alias{annotate_plates}
\title{RStudio Addin for annotating plate images}
\usage{
annotate_plates(dir = file.choose(), n_positions = 9L, queries = NULL,
  strain_collections = NULL, media = NULL, treatments = NULL,
  temperatures = c(23, 27, 30, 33, 37))
}
\arguments{
\item{dir}{The directory containing images to annotate. Default setting
prompts the user to choose a file.}

\item{n_positions}{The default number of plate positions per image
(defaults to \code{9L}).}

\item{queries}{An optional vector of available query IDs.}

\item{strain_collections}{An optional vector of available strain collection IDs.}

\item{media}{An optional vector of available media IDs.}

\item{treatments}{An optional vector of available treatment IDs.}

\item{temperatures}{A vector of recommended temperatures.}
}
\description{
This function starts a Shiny gadget that allows a user to record information
about plate images for arrayed colony growth experiments.
}
\details{
This application is provided to make annotation of plates in
colony growth experiments easier and safer. It will try to warn, or prevent
the user from entering invalid data. If the \code{rothfreezer} package
is installed, arguments set to \code{NULL} will be pulled from this package's
database.

The resulting annotation table is saved as \code{plate-annotations.csv}
in the directory containing the annotated images.

To quit without saving, just press "Cancel".
}

