% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_crop}
\alias{calibrate_crop}
\title{Calibrate cropping and rotation parameters}
\usage{
calibrate_crop(dir, rotate = 90, range = 6, step = 0.2, thresh = 0.03,
  invert = TRUE, rough_pad = c(0, 0, 0, 0), fine_pad = c(5, 5, 5, 5),
  display = TRUE, overwrite = FALSE)
}
\arguments{
\item{dir}{Directory of images to process.}

\item{rotate}{A rough angle in degrees clockwise to rotate each plate. The
rotation angle will be further calibrated after applying this rotation.
Defaults to \code{90}.}

\item{range}{Range to explore (in degrees) when calibrating rotation angle.
Defaults to \code{6}.}

\item{step}{Increment (in degrees) to step when calibrating rotation angle.
Defaults to \code{0.2}.}

\item{thresh}{Fraction of foreground pixels needed to identify plate
boundaries when rough cropping. Defaults to \code{0.03}.}

\item{invert}{Should the image be inverted? Defaults to \code{TRUE}.
Recommended \code{TRUE} if colonies are darker than the plate.}

\item{display}{Should cropped images be displayed for review?
Defaults to \code{TRUE}.}

\item{overwrite}{Should existing crop calibration be overwritten?
Defaults to \code{FALSE}.}
}
\description{
This function calibrates plate cropping and rotation parameters for an image
with an arbritrarily sized grid of plates.
}
\details{
Crop calibration procedes through the following 3 steps:

\enumerate{
  \item Rough crop
  \item Rotate
  \item Fine crop
}

Rough cropping relies on high contrast between plates. If
\code{invert = TRUE} plates should be light and the region between plates
should be dark, and vice versa if \code{invert = FALSE}.

Rotation first applies the \code{rotate} argument, then iterates through
a range of degrees specified by the \code{range} argument with a step
specified by the \code{step} argument. The image is first thresholded to identify
objects (i.e. colonies) and the objects are expanded to get a rough shape of
their location on the plate. The final rotation angle is chosen
by minimizing the variance of rowwise sums for the range of rotation angles
explored, which effectively aligns a rectangular shape with the axes.

Fine cropping finds the nearest object edge (problematic for plates without
any growth on the intended grid edges).
}

