% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_grid}
\alias{calibrate_grid}
\title{Calibrate Colony Grid}
\usage{
calibrate_grid(dir, invert = TRUE, display = TRUE, overwrite = FALSE)
}
\arguments{
\item{dir}{Path to screenmill project directory.}

\item{invert}{Should the image be inverted. (Recommended \code{TRUE} if
colonies are darker than the background).}

\item{display}{Should calibrated grid be displayed for review?
Defaults to \code{TRUE}.}
}
\description{
Calibrates colony grid coordinates in for each unique grid template found
in \code{screenmill-plates.csv} (see \link{annotate_plates}).
}

