% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_gather.R
\name{create_key_from_csv_tables}
\alias{create_key_from_csv_tables}
\title{Gather annotation information in row by column format and convert into long format.}
\usage{
create_key_from_csv_tables(dir, strain_collection_id)
}
\arguments{
\item{dir}{directory containing strain annotation csv files}

\item{strain_collection_id}{string containing collection id}

\item{path}{Path to CSV to gather}

\item{value}{Name of value variable}

\item{plate_pattern}{Regex used to extract plate number from `path`}
}
\description{
Given a table with columns "row", "1", "2", ... gather everything but the "row"
column into a table with variables "row", "column", `value`, "plate".
}
\details{
It is easiest to annotate e.g. a 96 well plate with strain, plasmid,
OD600 etc. information in a way that looks visually similar to the plate.
However this results in multiple tables (one for each annotated variable) in
wide format. This function will convert such tables into long format such
that each variable is in a single column rather than spread out in the grid
style plate layout. Multiple gathered annotations can easily be joined to
a single table
}
