% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-crop.R
\name{crop}
\alias{crop}
\title{Crop plates}
\usage{
crop(screenmill_plates = "test-scans/screenmill-plates.csv")
}
\arguments{
\item{screenmill_plates}{Path to a screenmill plates annotation CSV file
(the result of \link{annotate_plates}).}
}
\description{
Crop plates from images.
}

