% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{display_calibration}
\alias{display_calibration}
\title{Display crop calibration}
\usage{
display_calibration(dir = ".", groups = NULL, positions = NULL)
}
\arguments{
\item{dir}{Directory of images}

\item{groups}{Cropping groups to display. Defaults to \code{NULL} which will
display all groups.}

\item{positions}{Positions to display. Defaults to \code{NULL} which will
display all positions.}
}
\description{
Convenience function for displaying crop calibrations. Usefull for viewing
the result of manually edited
}

