% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-engine.R
\name{fine_crop}
\alias{fine_crop}
\title{Fine crop plates}
\usage{
fine_crop(img, rotate, range, incr, pad, invert)
}
\arguments{
\item{img}{Image}

\item{rotate}{Rotate}

\item{pad}{Pad}

\item{invert}{Invert}
}
\description{
Fine crop plates
}

