% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-engine.R
\name{grid_angle}
\alias{grid_angle}
\title{Determine plate rotation angle}
\usage{
grid_angle(img, rough, range, step)
}
\arguments{
\item{img}{A rough cropped plate image}

\item{rough}{Rough angle in degrees clockwise with which to rotate plate}

\item{step}{Degree increment used in fine rotation}
}
\description{
Uses a Radon transform to determine the angle that aligns the colony grid
}
\details{
The implementation of this function was adapted from Gitter
(see \href{http://www.ncbi.nlm.nih.gov/pubmed/24474170}{PMID:24474170})
}

