% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-grid-breaks.R
\name{grid_breaks}
\alias{grid_breaks}
\title{Determine column or row breaks for grid}
\usage{
grid_breaks(img, type = c("col", "row"), thresh = 0.01, edges = c("inner",
  "outer", "mid"))
}
\arguments{
\item{img}{An Image object or a matrix}

\item{type}{Type of grid breaks ('column' or 'row') defaults to column.}

\item{thresh}{Threshold used to define local valleys in image.}

\item{edges}{How to handle edge breaks. Defaults to 'inner' which will use
the inner edge of flanking valleys. 'outer' will use outer edge of flanking
valleys. Otherwise the midpoint of flanking valleys will be returned.}
}
\description{
Uses midpoints of low average pixel intensity to determine column breaks.
}

