% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth-parameters.R
\name{growth_parameters}
\alias{growth_parameters}
\title{Calculate growth parameters}
\usage{
growth_parameters(time, size, ...)
}
\arguments{
\item{time}{Numeric - time of growth.}

\item{size}{Numeric - size of colony.}

\item{...}{Further arguments passed to \code{\link[stats]{smooth.spline}}.}
}
\value{
Returns a data frame with the following parameters derived from the
fit curve:
\tabular{ll}{
  \bold{A}        \tab Maximum growth. \cr
  \bold{A_t}      \tab Time at maximum growth. \cr
  \bold{mu}       \tab Maximum growth rate. \cr
  \bold{mu_t}     \tab Time at maximum growth rate. \cr
  \bold{mu_y}     \tab Growth at maximum growth rate. \cr
  \bold{lambda}   \tab Lag phase (x-intercept of line tangent to max growth
                       rate). \cr
  \bold{b}        \tab y-intercept of line tangent to max growth rate. \cr
  \bold{integral} \tab Area under growth curve. \cr
  \bold{spar}     \tab Smoothing parameter used to generate curve - can be
                       set by passing a \code{spar} argument to
                       \code{\link[stats]{smooth.spline}}.
}
}
\description{
Calculates colony growth parameters by fitting a smooth growth curve using
\code{\link[stats]{smooth.spline}}.
}

