% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Measure Colonies}
\usage{
measure(dir = ".", overwrite = F, save.plates = F, save.colonies = T)
}
\arguments{
\item{dir}{Path to directory to process.}

\item{overwrite}{Should previous measurements be overwritten?}

\item{save.plates}{Should cropped/rotated plates be saved? Defaults to \code{FALSE}.}

\item{save.colonies}{Should cropped colonies be saved? Defaults to \code{TRUE}.}
}
\description{
Measure and save colonies.
}

