% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-templates.R
\name{new_batch_metadata}
\alias{new_batch_metadata}
\title{Write metadata templates}
\usage{
new_batch_metadata(cm)
}
\arguments{
\item{cm}{Path to colony measurements (CM) file.}
}
\description{
A convenience function to create a templated 'screens.csv'
and 'plates.csv' metadata file for a batch of screens.
}

