% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-templates.R
\name{new_strain_collection}
\alias{new_strain_collection}
\title{Create new strain collection}
\usage{
new_strain_collection(id, nplates, format = 96, dim = c(2, 3))
}
\arguments{
\item{id}{ID of new library. This ID needs to be unique in the database.}

\item{nplates}{Number of plates in the library.}

\item{format}{Size of plates. Defaults to 96.}

\item{dim}{Aspect ratio of rows to columns. Defaults to \code{c(2, 3)}.}
}
\description{
A convenience function to create a new strain collection keyfile
}

