% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-plates.R
\name{plot_plates}
\alias{plot_plates}
\title{Plot plate as heatmap}
\usage{
plot_plates(data, fill, row, col, plate = "plate", query = "query_name",
  upper = 1, mid = 0, lower = -1)
}
\arguments{
\item{data}{A data frame}

\item{fill}{String. Name of fill column.}

\item{row}{String. Name of row column.}

\item{col}{String. Name of column column.}

\item{plate}{String. Name of plate column. Defaults to 'plate'.}

\item{query}{String. Name of query column. Defaults to 'query_name'.}

\item{upper}{Scalar. Upper limit. Defaults to 1.}

\item{mid}{Scalar. Middle of scale. Defaults to 0.}

\item{lower}{Scalar. Lower limit. Defaults to -1.}
}
\value{
A ggplot object
}
\description{
This function will plot plates ~ queries for a desired metric with forced
upper and lower bounds
}

