% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cm}
\alias{read_cm}
\title{Read CM engine log file.}
\usage{
read_cm(path, replicates, dim = c(2, 3))
}
\arguments{
\item{path}{Path to a CM engine log file.}

\item{replicates}{Integer. The number of replicates for each strain.}

\item{dim}{Numeric vector of length 2. The aspect ratio of the colonies on a
plate. Defaults to 2 rows for every 3 columns.}
}
\value{
A dataframe with columns:
\itemize{
  \item{\bold{scan_name:}} The name of the scanned plate.
  \item{\bold{scan_cond:}} The condition of the scanned plate.
  \item{\bold{plate:}} The strain library plate number.
  \item{\bold{row:}} The strain library row.
  \item{\bold{column:}} The strain library column.
  \item{\bold{replicate:}} The strain replicate number.
  \item{\bold{size:}} The size of the colony (unadjusted).
  \item{\bold{circ:}} The circularity of the colony.
}
}
\description{
Reads ScreenMill Colony Measurement engine log file.
}

