% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_dr}
\alias{read_dr}
\title{Read DR stats file.}
\usage{
read_dr(path, match = "Query\\tCondition\\tPlate #\\tRow\\tColumn")
}
\arguments{
\item{path}{Path to a DR stats file. Defaults to interactive file selection.}

\item{match}{regex used to find the line containing the table header.}
}
\value{
A dataframe with columns:
\itemize{
  \item{\bold{scan_name:}} The name of the scanned plate.
  \item{\bold{scan_cond:}} The condition of the scanned plate.
  \item{\bold{plate:}} The strain library plate number.
  \item{\bold{row:}} The strain library row.
  \item{\bold{column:}} The strain library column.
  \item{\bold{replicate:}} The strain replicate number.
  \item{\bold{size_dr:}} Size returned by DR engine.
  \item{\bold{excluded_query:}} Whether this observation was excluded in
  review.
  \item{\bold{excluded_control:}} Whether this observation's control was excluded
  in review.
}
}
\description{
Reads ScreenMill Data Review tables.
}

