% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_metadata}
\alias{read_metadata}
\title{Read rothscreen metadata}
\usage{
read_metadata(screens, plates)
}
\arguments{
\item{screens}{Path to screens metadata CSV}

\item{plates}{Path to plates metadata CSV}
}
\description{
A convenience function to read and combine rothscreen metadata
}

