% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_plate_layout}
\alias{read_plate_layout}
\title{Read plate layout

Reads a directory of plate layout CSVs exported from Numbers (for mac) into a dataframe.}
\usage{
read_plate_layout(dir)
}
\arguments{
\item{dir}{A directory.}
}
\description{
Read plate layout

Reads a directory of plate layout CSVs exported from Numbers (for mac) into a dataframe.
}
\details{
To generate a plate layout directory with Numbers, make a separate sheet for each plate
and name each sheet "plate\verb{<number>}" (Replace \verb{<number>} with the plate number). Then create one
table for each annotation with columns named "row", "1", "2", "3" etc. for the dimension of the
plate. In the first "row" column, number each row "1", "2", "3" etc. for the dimension of the
plate. Name each tabel with the intended annotation (e.g. "strain_id" for a table annotating
strain IDs for each position in the plate). Once you have annotated your plates, you can export
them to CSV (\verb{File>Export To>CSV...}). Name the resulting directory with the desired strain
collection ID (some combination of the date and a descriptive word is recommended). The files
in this directory will be named \verb{<sheet>-<table>.csv} (e.g. \code{plate1-strain_id.csv}).

These tables are required for every plate:
\itemize{
\item \strong{strain_id} - The strain identifier.
\item \strong{strain_name} - The human-readable name of the strain.
\item \strong{plate_control} - (T/F) is position intended to be used as a control for plate normalization?
\item \emph{excluded} - positions to exclude from analysis.
}

These tables are recommended (depending on the experiment):
\itemize{
\item \emph{plasmid_id} - A plasmid identifier
\item \emph{OD600} - The culture density used for pinning.
\item \emph{pinnings} - The number of pinnings from source plate to target plate.
\item \emph{start_temp} - The starting temperature at the beginning of the experiment.
\item \emph{end_temp} - The ending temperature at the end of the experiment.
\item \emph{start_humidity} - The starting humidity at the beginning of the experiment.
\item \emph{end_humidity} - The ending humidity at the end of the experiment.
}
}
