% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{Interactive colony review tool}
\usage{
review(dir = ".", overwrite = FALSE)
}
\arguments{
\item{dir}{Directory of images to process.}

\item{overwrite}{Should existing crop calibration be overwritten?
Defaults to \code{FALSE}.}
}
\description{
Manually exclude regions of plates with contamination or obvious mispinnings.
}
