% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-rough-crop.R
\name{rough_crop}
\alias{rough_crop}
\title{Rough crop a grid of plates}
\usage{
rough_crop(img, thresh = 0.5, dim = c(3, 3), pad = c(0, 0, 0, 0),
  display = FALSE)
}
\arguments{
\item{img}{Either a path to a plate grid image file, or its "Image" object.
(see \link[EBImage]{readImage})}

\item{thresh}{Pixel intensity threshold used to find plate edges. Defaults
to \code{0.5}. Note: region between plates is assumed to be dark.}

\item{dim}{Grid dimensions. Defaults to \code{c(3, 3)} for three rows and
three columns of plates respectively.}

\item{pad}{Number of pixels to add (or remove) from detected plate edges.
Defaults to \code{c(0, 0, 0, 0)} which adds 0 pixels to the
left, right, top, and bottom coordinates of each plate respectively. Negative
values will remove pixels.}

\item{display}{Should the resulting cropped images be displayed in a web
browser? Defaults to \code{FALSE}.}
}
\value{
\code{rough_crop} returns a dataframe with the following columns:

\item{position}{Integer position of plate in grid (row-wise).}
\item{rough_top}{The rough top edge of the plate.}
\item{rough_left}{The rough left edge of the plate.}
\item{rough_right}{The rough right edge of the plate.}
\item{rough_bot}{The rough bottom edge of the plate.}
}
\description{
Finds the rough plate edges for a grid of plates.
}
\details{
Rough crops for a grid of plates are detected by thresholding by the
brightness of the image (can be adjusted using the \code{thresh} argument).
Rows and columns are then scanned to identify locations with a large number
of pixels that are more than the threshold pixel intensity. These transitions
are used to define the plate edges. For best results, the grid should be
approximately square to the edge of the image. Plate positions are numbered
from left to right, top to bottom.
}

