% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-screenmill.R
\docType{package}
\name{screenmill}
\alias{screenmill}
\alias{screenmill-package}
\title{Screenmill R package}
\description{
Screenmill R package
}
\author{
Eric Bryant \email{eeb2139@columbia.edu},

Maintainer: Eric Bryant \email{eeb2139@columbia.edu}
}

