% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-effect.R
\name{spatial_effect}
\alias{spatial_effect}
\title{Compute spatial effect}
\usage{
spatial_effect(x, y, value, method = c("rlm", "loess"), deg = 2, ...)
}
\arguments{
\item{x}{Integer vector - x colony position (e.g. column).}

\item{y}{Integer vector - y colony position (e.g. row).}

\item{value}{Numeric vector - measured colony size.}

\item{method}{Method used for fitting smooth surface. Either \code{'rlm'},
or \code{'loess'}. Defaults to \code{'rlm'}.}

\item{deg}{Degree of polynomial used for fitting. See \link[stats]{poly} for
details. Defaults to 2.}

\item{...}{Further arguments passed to method.}
}
\value{
Returns a vector with the same length as \code{value}.
}
\description{
Computes the effect of plate position on colony growth.
}
\seealso{
\code{\link[MASS]{rlm}} \code{\link[stats]{loess}}
}

