% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-crop.R
\name{crop}
\alias{crop}
\title{Crop plates}
\usage{
crop(dir, target = "cropped", overwrite = FALSE)
}
\arguments{
\item{dir}{Directory of images to crop.}

\item{target}{Target directory for saving images.}

\item{overwrite}{Should existing cropped images be overwritten?
Defaults to \code{FALSE}.}
}
\description{
Crop plates from images.
}

