% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_screenmill}
\alias{read_screenmill}
\title{Read screenmill files in a directory}
\usage{
read_screenmill(dir)
}
\arguments{
\item{dir}{Directory containing screenmill files}
}
\description{
Read screenmill files in a directory
}
