% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-rough-crop.R
\name{rough_crop}
\alias{rough_crop}
\title{Rough crop a grid of plates}
\usage{
rough_crop(path, thresh = 0.5, dim = c(3, 3), pad = c(-20, -20, -20, -20),
  display = FALSE)
}
\arguments{
\item{path}{Path to an image of a plate grid.}

\item{thresh}{Pixel intensity threshold used to find plate edges. Defaults
to \code{0.5}. Note: region between plates is assumed to be dark.}

\item{dim}{Grid dimensions. Defaults to \code{c(3, 3)} for three rows and
three columns of plates respectively.}

\item{pad}{Number of pixels to add (or remove) from detected plate edges.
Defaults to \code{c(-20, -20, -20, -20)} which removes 20 pixels from the
left, right, top, and bottom of each plate respectively.}

\item{display}{Should the resulting cropped images be displayed in a web
browser? Defaults to \code{FALSE}.}
}
\value{
\code{rough_crop} returns a dataframe with the following columns:

\item{plate}{Integer position of plate in grid (row-wise).}
\item{rough_top}{The rough top edge of the plate.}
\item{rough_left}{The rough left edge of the plate.}
\item{rough_right}{The rough right edge of the plate.}
\item{rough_bot}{The rough bottom edge of the plate.}
}
\description{
Finds and returns the rough plate edges for a grid of plates.
}
\details{
Rough crops for a grid of plates are detected by thresholding by the
brightness of the image (can be adjusted using the \code{thresh} argument).
Rows and columns are then scanned to identify locations with a large number
of pixels that are more than the threshold pixel intensity. These transitions
are used to define the plate edges. For best results, the grid should be
approximately square to the edge of the image. Plates are numbered from
left to right, top to bottom.
}

