% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-rough-crop.R
\name{rough_crop}
\alias{rough_crop}
\title{Rough crop a grid of plates}
\usage{
rough_crop(img, thresh, invert, pad)
}
\arguments{
\item{img}{Either an "Image" object (see \link[EBImage]{readImage}), or a
matrix.}

\item{thresh}{Fraction of foreground pixels needed to identify plate
boundaries. Defaults to \code{0.03}.}

\item{invert}{Should the image be inverted? Defaults to \code{TRUE}.
Recommended \code{FALSE} if region between plates is lighter than the
plates. This is faster than manual inversion of entire image.}

\item{pad}{Number of pixels to add (or remove) from detected plate edges.
Defaults to \code{c(0, 0, 0, 0)} which adds 0 pixels to the
left, right, top, and bottom coordinates of each plate respectively. Negative
values will remove pixels.}

\item{display}{Should the resulting cropped images be displayed?
Defaults to \code{FALSE}.}
}
\value{
\code{rough_crop} returns a dataframe with the following columns:

\item{position}{Integer position of plate in grid (row-wise).}
\item{top}{The rough top edge of the plate.}
\item{left}{The rough left edge of the plate.}
\item{right}{The rough right edge of the plate.}
\item{bot}{The rough bottom edge of the plate.}
}
\description{
Finds the rough plate edges for a grid of plates.
}
\details{
Rough crops for a grid of plates are detected by thresholding by the
brightness of the image (can be adjusted using the \code{thresh} argument).
Rows and columns are then scanned to identify locations with a large number
of pixels that are more than the threshold pixel intensity. These transitions
are used to define the plate edges. For best results, the grid should be
approximately square to the edge of the image. Plate positions are numbered
from left to right, top to bottom.
}

